+++ example/src/Screens/BottomTabs.tsx
@@ -5,10 +5,16 @@ import {
   TransitionPresets,
   useBottomTabBarHeight,
 } from '@react-navigation/bottom-tabs';
-import { HeaderBackButton, useHeaderHeight } from '@react-navigation/elements';
+import {
+  HeaderBackButton,
+  HeaderButton,
+  PlatformPressable,
+  useHeaderHeight,
+} from '@react-navigation/elements';
 import {
   type NavigatorScreenParams,
   type ParamListBase,
+  type PathConfigMap,
   useIsFocused,
 } from '@react-navigation/native';
 import type { StackScreenProps } from '@react-navigation/stack';
@@ -22,12 +28,15 @@ import {
   StyleSheet,
   useWindowDimensions,
 } from 'react-native';
-import { Appbar, IconButton } from 'react-native-paper';
 
 import { Albums } from '../Shared/Albums';
 import { Chat } from '../Shared/Chat';
 import { Contacts } from '../Shared/Contacts';
-import { SimpleStack, type SimpleStackParams } from './SimpleStack';
+import {
+  SimpleStack,
+  simpleStackLinking,
+  type SimpleStackParams,
+} from './SimpleStack';
 
 const getTabBarIcon =
   (name: React.ComponentProps<typeof MaterialCommunityIcons>['name']) =>
@@ -35,11 +44,21 @@ const getTabBarIcon =
     <MaterialCommunityIcons name={name} color={color} size={size} />
   );
 
-type BottomTabParams = {
+export type BottomTabParams = {
   TabStack: NavigatorScreenParams<SimpleStackParams>;
   TabAlbums: undefined;
   TabContacts: undefined;
   TabChat: undefined;
+};
+
+export const bottomTabLinking: PathConfigMap<BottomTabParams> = {
+  TabStack: {
+    path: 'stack',
+    screens: simpleStackLinking,
+  },
+  TabAlbums: 'albums',
+  TabContacts: 'contacts',
+  TabChat: 'chat',
 };
 
 const AlbumsScreen = () => {
@@ -99,9 +118,7 @@ export function BottomTabs({
             <HeaderBackButton {...props} onPress={navigation.goBack} />
           ),
           headerRight: ({ tintColor }) => (
-            <Appbar.Action
-              icon={animation === 'none' ? 'heart-outline' : 'heart'}
-              color={tintColor}
+            <HeaderButton
               onPress={() => {
                 const options = Object.keys(
                   animations
@@ -124,7 +141,13 @@ export function BottomTabs({
                   }
                 );
               }}
-            />
+            >
+              <MaterialCommunityIcons
+                name={animation === 'none' ? 'heart-outline' : 'heart'}
+                size={24}
+                color={tintColor}
+              />
+            </HeaderButton>
           ),
           tabBarPosition: isLargeScreen ? 'left' : 'bottom',
           tabBarLabelPosition:
@@ -209,15 +232,21 @@ export function BottomTabs({
         />
       </Tab.Navigator>
       {isLargeScreen ? (
-        <IconButton
-          icon={isCompact ? 'chevron-double-right' : 'chevron-double-left'}
+        <PlatformPressable
           onPress={() => setIsCompact(!isCompact)}
           style={{
             position: 'absolute',
             bottom: 0,
             left: 0,
+            padding: 16,
           }}
-        />
+        >
+          <MaterialCommunityIcons
+            name={isCompact ? 'chevron-double-right' : 'chevron-double-left'}
+            size={24}
+            color="black"
+          />
+        </PlatformPressable>
       ) : null}
     </>
   );