+++ packages/core/src/__tests__/usePreventRemove.test.tsx
@@ -10,7 +10,7 @@ import { BaseNavigationContainer } from '../BaseNavigationContainer';
 import { createNavigationContainerRef } from '../createNavigationContainerRef';
 import { Screen } from '../Screen';
 import { useNavigationBuilder } from '../useNavigationBuilder';
-import { UNSTABLE_usePreventRemove as usePreventRemove } from '../usePreventRemove';
+import { usePreventRemove } from '../usePreventRemove';
 import { MockRouterKey } from './__fixtures__/MockRouter';
 
 jest.mock('nanoid/non-secure', () => {
@@ -948,16 +948,15 @@ it("prevents removing a child screen with 'usePreventRemove' hook with 'resetRoo
     type: 'stack',
   });
 
-  act(
-    () =>
-      ref.current?.resetRoot({
-        index: 0,
-        key: 'stack-2',
-        routeNames: ['foo', 'bar', 'baz'],
-        routes: [{ key: 'foo-3', name: 'foo' }],
-        stale: false,
-        type: 'stack',
-      })
+  act(() =>
+    ref.current?.resetRoot({
+      index: 0,
+      key: 'stack-2',
+      routeNames: ['foo', 'bar', 'baz'],
+      routes: [{ key: 'foo-3', name: 'foo' }],
+      stale: false,
+      type: 'stack',
+    })
   );
 
   expect(onStateChange).toHaveBeenCalledTimes(1);