+++ packages/devtools/src/parseErrorStack.tsx
@@ -49,11 +49,11 @@ export function parseErrorStack(
   const parsedStack = Array.isArray(errorStack)
     ? errorStack
     : (global as any).HermesInternal
-    ? convertHermesStack(parseHermesStack(errorStack))
-    : stacktraceParser.parse(errorStack).map((frame) => ({
-        ...frame,
-        column: frame.column != null ? frame.column - 1 : null,
-      }));
+      ? convertHermesStack(parseHermesStack(errorStack))
+      : stacktraceParser.parse(errorStack).map((frame) => ({
+          ...frame,
+          column: frame.column != null ? frame.column - 1 : null,
+        }));
 
   return parsedStack as StackFrame[];
 }