+++ packages/devtools/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [7.0.0-alpha.7](https://github.com/react-navigation/react-navigation/compare/@react-navigation/devtools@7.0.0-alpha.6...@react-navigation/devtools@7.0.0-alpha.7) (2024-01-17)
+
+**Note:** Version bump only for package @react-navigation/devtools
+
 # [7.0.0-alpha.6](https://github.com/react-navigation/react-navigation/compare/@react-navigation/devtools@7.0.0-alpha.5...@react-navigation/devtools@7.0.0-alpha.6) (2023-11-17)
 
 **Note:** Version bump only for package @react-navigation/devtools