+++ packages/stack/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [7.0.0-alpha.8](https://github.com/react-navigation/react-navigation/compare/@react-navigation/stack@7.0.0-alpha.7...@react-navigation/stack@7.0.0-alpha.8) (2024-01-17)
+
+### Bug Fixes
+
+* remove optional check in CardContainer ([#11772](https://github.com/react-navigation/react-navigation/issues/11772)) ([e73551a](https://github.com/react-navigation/react-navigation/commit/e73551a237f36e3242b329b71322e22d0190af74)) - by @osdnk
+* remove unnecessary check in the stack navigator ([#11747](https://github.com/react-navigation/react-navigation/issues/11747)) ([02ec21b](https://github.com/react-navigation/react-navigation/commit/02ec21b0b5660ff17758b596852a7d7dc06ff69e)), closes [#11733](https://github.com/react-navigation/react-navigation/issues/11733) - by @osdnk
+
+### Features
+
+* add layout and screenLayout props for screens ([#11741](https://github.com/react-navigation/react-navigation/issues/11741)) ([2dc2178](https://github.com/react-navigation/react-navigation/commit/2dc217827a1caa615460563973d3d658be372b29)) - by @satya164
+* add retaining of the screen to the stack navigator ([#11765](https://github.com/react-navigation/react-navigation/issues/11765)) ([7fe82f7](https://github.com/react-navigation/react-navigation/commit/7fe82f7217d9d146e5a127bcac921f0c90e2059f)), closes [#11758](https://github.com/react-navigation/react-navigation/issues/11758) - by @osdnk
+* preloading for stack navigator ([#11733](https://github.com/react-navigation/react-navigation/issues/11733)) ([14fa6df](https://github.com/react-navigation/react-navigation/commit/14fa6dfa4484cf2784f0e5cd0d06252fdf8a4ba5)), closes [#11702](https://github.com/react-navigation/react-navigation/issues/11702) [#11727](https://github.com/react-navigation/react-navigation/issues/11727) - by @osdnk
+* preloading in routers  ([382d6e6](https://github.com/react-navigation/react-navigation/commit/382d6e6f3312630b34332b1ae7d4bd7bf9b4ee60)) - by @osdnk
+
 # [7.0.0-alpha.7](https://github.com/react-navigation/react-navigation/compare/@react-navigation/stack@7.0.0-alpha.6...@react-navigation/stack@7.0.0-alpha.7) (2023-11-17)
 
 ### Bug Fixes